modify all modules
