--- consumer/src/main/java/me/danilepf/reactivemessagingplay/consumer/ConsumerApplication.java
@@ -1,13 +1,0 @@
-package me.danilepf.reactivemessagingplay.consumer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ConsumerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ConsumerApplication.class, args);
-    }
-
-}
