--- notifier/src/main/java/me/danilepf/reactivemessagingplay/notifier/NotifierApplication.java
@@ -1,13 +1,0 @@
-package me.danilepf.reactivemessagingplay.notifier;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotifierApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NotifierApplication.class, args);
-    }
-
-}
