--- producer/src/main/java/me/danilepf/reactivemessagingplay/producer/ProducerApplication.java
@@ -1,13 +1,0 @@
-package me.danilepf.reactivemessagingplay.producer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProducerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProducerApplication.class, args);
-    }
-
-}
