+++ rmp-consumer/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+server.port=8082
