+++ rmp-notifier/src/main/java/me/danielpf/reactivemessagingplay/notifier/NotifierApplication.java
@@ -1,0 +1,13 @@
+package me.danielpf.reactivemessagingplay.notifier;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NotifierApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(NotifierApplication.class, args);
+    }
+
+}
