+++ rmp-notifier/src/main/resources/application.properties
