+++ rmp-producer/src/main/java/me/danielpf/reactivemessagingplay/producer/ProducerApplication.java
@@ -1,0 +1,13 @@
+package me.danielpf.reactivemessagingplay.producer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProducerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProducerApplication.class, args);
+    }
+
+}
